refactored dolfinx spin example
