+++ utopia/external/fenics/dolfinx_spin/main.cpp
@@ -18,192 +18,27 @@
 #include "utopia_TwoFieldAlternateMinimization.hpp"
 #include "utopia_TwoFieldSPIN.hpp"
 
+#include "DolfinxFunction.hpp"
+
 using namespace dolfinx;
 using T = PetscScalar;
 
-class HyperElasticProblem : public utopia::Function<utopia::PetscMatrix, utopia::PetscVector> {
-public:
-    class Impl {
-    public:
-        Impl(std::shared_ptr<fem::FunctionSpace> V,
-             std::shared_ptr<fem::Function<T>> u,
-             std::shared_ptr<fem::Form<T>> objective,
-             std::shared_ptr<fem::Form<T>> gradient,
-             std::shared_ptr<fem::Form<T>> hessian,
-             std::vector<std::shared_ptr<const fem::DirichletBC<T>>> boundary_conditions)
-            : V(V),
-              u(u),
-              objective(objective),
-              gradient(gradient),
-              hessian(hessian),
-              boundary_conditions(boundary_conditions),
-              rhs(gradient->function_spaces()[0]->dofmap()->index_map,
-                  gradient->function_spaces()[0]->dofmap()->index_map_bs()),
-              matrix(la::petsc::Matrix(fem::petsc::create_matrix(*hessian, "baij"), false)),
-              solution_vector(la::petsc::Vector(la::petsc::create_vector_wrap(*u->x()), false)) {
-            auto map = gradient->function_spaces()[0]->dofmap()->index_map;
-            const int bs = gradient->function_spaces()[0]->dofmap()->index_map_bs();
-            std::int32_t size_local = bs * map->size_local();
+std::shared_ptr<DolfinxFunction> hyperelasticity(const std::shared_ptr<fem::FunctionSpace> &V,
+                        std::vector<std::shared_ptr<const fem::DirichletBC<T>>> boundary_conditions)
+{
+    auto u = std::make_shared<fem::Function<T>>(V);
 
-            std::vector<PetscInt> ghosts(map->ghosts().begin(), map->ghosts().end());
-            std::int64_t size_global = bs * map->size_global();
-            VecCreateGhostBlockWithArray(map->comm(),
-                                         bs,
-                                         size_local,
-                                         size_global,
-                                         ghosts.size(),
-                                         ghosts.data(),
-                                         rhs.array().data(),
-                                         &rhs_petsc_);
-        }
+    auto objective = std::make_shared<fem::Form<T>>(
+        fem::create_form<T>(*form_hyperelasticity_Pi, {}, {{"u", u}}, {}, {}, V->mesh()));
 
-        ~Impl() {
-            if (rhs_petsc_) VecDestroy(&rhs_petsc_);
-        }
+    auto gradient =
+        std::make_shared<fem::Form<T>>(fem::create_form<T>(*form_hyperelasticity_F_form, {V}, {{"u", u}}, {}, {}));
 
-        std::shared_ptr<fem::FunctionSpace> V;
-        std::shared_ptr<fem::Function<T>> u;
-        std::shared_ptr<fem::Form<T>> objective, gradient, hessian;
-        std::vector<std::shared_ptr<const fem::DirichletBC<T>>> boundary_conditions;
+    auto hessian = std::make_shared<fem::Form<T>>(
+        fem::create_form<T>(*form_hyperelasticity_J_form, {V, V}, {{"u", u}}, {}, {}));
 
-        la::Vector<T> rhs;
-        la::petsc::Matrix matrix;
-        la::petsc::Vector solution_vector;
-
-        Vec rhs_petsc_;
-
-        auto form() {
-            return [](Vec x) {
-                VecGhostUpdateBegin(x, INSERT_VALUES, SCATTER_FORWARD);
-                VecGhostUpdateEnd(x, INSERT_VALUES, SCATTER_FORWARD);
-            };
-        }
-
-        /// Compute F at current point x
-        auto F() {
-            return [&](const Vec x, Vec) {
-                // Assemble b and update ghosts
-                xtl::span<T> b(rhs.mutable_array());
-                std::fill(b.begin(), b.end(), 0.0);
-                fem::assemble_vector<T>(b, *gradient);
-                VecGhostUpdateBegin(rhs_petsc_, ADD_VALUES, SCATTER_REVERSE);
-                VecGhostUpdateEnd(rhs_petsc_, ADD_VALUES, SCATTER_REVERSE);
-
-                // Set bcs
-                Vec x_local;
-                VecGhostGetLocalForm(x, &x_local);
-                PetscInt n = 0;
-                VecGetSize(x_local, &n);
-                const T *array = nullptr;
-                VecGetArrayRead(x_local, &array);
-
-                // FIXME Boundary conditions should be set to if x satisfies them (!!!)
-                fem::set_bc<T>(b, boundary_conditions, xtl::span<const T>(array, n), -1.0);
-                VecRestoreArrayRead(x, &array);
-            };
-        }
-
-        /// Compute J = F' at current point x
-        auto J() {
-            return [&](const Vec, Mat A) {
-                MatZeroEntries(A);
-                fem::assemble_matrix(la::petsc::Matrix::set_block_fn(A, ADD_VALUES), *hessian, boundary_conditions);
-                MatAssemblyBegin(A, MAT_FLUSH_ASSEMBLY);
-                MatAssemblyEnd(A, MAT_FLUSH_ASSEMBLY);
-                fem::set_diagonal(
-                    la::petsc::Matrix::set_fn(A, INSERT_VALUES), *hessian->function_spaces()[0], boundary_conditions);
-                MatAssemblyBegin(A, MAT_FINAL_ASSEMBLY);
-                MatAssemblyEnd(A, MAT_FINAL_ASSEMBLY);
-            };
-        }
-    };
-
-    HyperElasticProblem(const std::shared_ptr<fem::FunctionSpace> &V,
-                        std::vector<std::shared_ptr<const fem::DirichletBC<T>>> boundary_conditions) {
-        auto u = std::make_shared<fem::Function<T>>(V);
-
-        auto objective = std::make_shared<fem::Form<T>>(
-            fem::create_form<T>(*form_hyperelasticity_Pi, {}, {{"u", u}}, {}, {}, V->mesh()));
-
-        auto gradient =
-            std::make_shared<fem::Form<T>>(fem::create_form<T>(*form_hyperelasticity_F_form, {V}, {{"u", u}}, {}, {}));
-
-        auto hessian = std::make_shared<fem::Form<T>>(
-            fem::create_form<T>(*form_hyperelasticity_J_form, {V, V}, {{"u", u}}, {}, {}));
-
-        impl_ = utopia::make_unique<Impl>(V, u, objective, gradient, hessian, boundary_conditions);
-    }
-
-    void create_vector(utopia::PetscVector &x) const override {
-        Vec v = impl_->solution_vector.vec();
-        x.wrap(v);
-    }
-
-    virtual ~HyperElasticProblem() {}
-
-    bool hessian(const utopia::PetscVector &x, utopia::PetscMatrix &H) const override {
-        UTOPIA_TRACE_SCOPE("HyperElasticProblem::hessian");
-
-        if (x.raw_type() != impl_->solution_vector.vec()) {
-            x.convert_to(impl_->solution_vector.vec());
-        }
-
-        auto H_fun = impl_->J();
-
-        if (H.empty()) {
-            auto m = impl_->matrix.mat();
-            H.wrap(m);
-        }
-
-        impl_->form()(impl_->solution_vector.vec());
-
-        H_fun(x.raw_type(), H.raw_type());
-        return true;
-    }
-
-    bool value(const utopia::PetscVector &x, PetscScalar &value) const override {
-        UTOPIA_TRACE_SCOPE("HyperElasticProblem::value");
-
-        if (x.raw_type() != impl_->solution_vector.vec()) {
-            x.convert_to(impl_->solution_vector.vec());
-        }
-
-        impl_->form()(impl_->solution_vector.vec());
-        value = fem::assemble_scalar<T>(*impl_->objective);
-        value = x.comm().sum(value);
-        return true;
-    }
-
-    bool gradient(const utopia::PetscVector &x, utopia::PetscVector &g) const override {
-        UTOPIA_TRACE_SCOPE("HyperElasticProblem::gradient");
-
-        if (x.raw_type() != impl_->solution_vector.vec()) {
-            x.convert_to(impl_->solution_vector.vec());
-        }
-
-        auto grad_fun = impl_->F();
-
-        impl_->form()(impl_->solution_vector.vec());
-
-        grad_fun(impl_->solution_vector.vec(), nullptr);
-
-        la::petsc::Vector temp(la::petsc::create_vector_wrap(impl_->rhs), false);
-
-        auto v = temp.vec();
-        assert(v);
-        g.copy_from(v);
-
-        if (g.has_nan_or_inf()) {
-            utopia::Utopia::Abort("HyperElasticProblem::gradient: Detected NaN!");
-        }
-        return true;
-    }
-
-    std::shared_ptr<fem::Function<T>> u() { return impl_->u; }
-
-private:
-    std::unique_ptr<Impl> impl_;
-};
+    return std::make_shared<DolfinxFunction>(V, u, objective, gradient, hessian, boundary_conditions);
+}
 
 int main(int argc, char *argv[]) {
     dolfinx::init_logging(argc, argv);
@@ -266,31 +101,16 @@ int main(int argc, char *argv[]) {
         auto bcs = std::vector{std::make_shared<const fem::DirichletBC<T>>(xt::xarray<T>{0, 0, 0}, bdofs_left, V),
                                std::make_shared<const fem::DirichletBC<T>>(u_rotation, bdofs_right)};
 
-        HyperElasticProblem problem(V, bcs);
-
         /////////////////////////////////////////////////////////
 
         bool verbose = true;
 
         // Split-fields
-        auto f1 = std::make_shared<HyperElasticProblem>(V, bcs);
-        auto f2 = std::make_shared<HyperElasticProblem>(V, bcs);
+        auto f1 = hyperelasticity(V, bcs);
+        auto f2 = hyperelasticity(V, bcs);
 
         // Global optimization problem (coupled fields)
-        auto c12 = std::make_shared<HyperElasticProblem>(V, bcs);
-
-        // auto nls1 = std::make_shared<utopia::Newton<utopia::PetscMatrix>>(
-        //     std::make_shared<utopia::Factorization<utopia::PetscMatrix, utopia::PetscVector>>());
-
-        // auto nls2 = std::make_shared<utopia::Newton<utopia::PetscMatrix>>(
-        //     std::make_shared<utopia::Factorization<utopia::PetscMatrix, utopia::PetscVector>>());
-
-        // auto nls1 = std::make_shared<utopia::Newton<utopia::PetscMatrix>>(
-        //     std::make_shared<utopia::ConjugateGradient<utopia::PetscMatrix, utopia::PetscVector, utopia::HOMEMADE>>());
-
-        // auto nls2 = std::make_shared<utopia::Newton<utopia::PetscMatrix>>(
-        //     std::make_shared<utopia::ConjugateGradient<utopia::PetscMatrix, utopia::PetscVector, utopia::HOMEMADE>>());
-
+        auto c12 = hyperelasticity(V, bcs);
 
         auto nls1 = std::make_shared<utopia::Newton<utopia::PetscMatrix>>(
             std::make_shared<utopia::BiCGStab<utopia::PetscMatrix, utopia::PetscVector, utopia::HOMEMADE>>());
@@ -321,13 +141,6 @@ int main(int argc, char *argv[]) {
             tfa.verbosity_level(utopia::VERBOSITY_LEVEL_DEBUG);
             tfa.solve(*c12, x);
         } else {
-            // auto lsc12 =
-            // std::make_shared<utopia::ConjugateGradient<utopia::PetscMatrix, utopia::PetscVector,
-            // utopia::HOMEMADE>>();
-            // lsc12->apply_gradient_descent_step(true);
-            // lsc12->set_preconditioner(std::make_shared<utopia::InvDiagPreconditioner<utopia::PetscMatrix,
-            // utopia::PetscVector>>());
-
             auto lsc12 = std::make_shared<utopia::KSP_MF<utopia::PetscMatrix, utopia::PetscVector>>();
             lsc12->ksp_type("gmres");
             lsc12->pc_type("none");
